Move JobAnalyticsCheck into impl
