+++ blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/analytics/FreeStyleAnalyticsCheck.java
@@ -3,7 +3,6 @@ package io.jenkins.blueocean.service.embedded.analytics;
 import hudson.Extension;
 import hudson.model.FreeStyleProject;
 import hudson.model.Item;
-import io.jenkins.blueocean.analytics.JobAnalyticsCheck;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
